Fix runtime bugs: currency symbols, missing @Transactional, NPE guards, stale closures
- Fix ₪ (shekel) → JOD in 8 backend spending limit messages (ClaimEngineService, GlobalPolicyClaimValidationService, PrescriptionService)
- Add @Transactional to SearchProfileService, PriceListService write methods, RoleService
- Add null checks for getMember()/getDoctor()/getClient() in EmergencyRequestService, LabRequestService, RadiologyRequestService, HealthcareProviderClaimService

+++ src/main/java/com/insurancesystem/Model/Entity/Claim.java
@@ -4,6 +4,7 @@ import com.insurancesystem.Model.Entity.Enums.ClaimStatus;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.util.UUID;
@@ -37,7 +38,7 @@ public class Claim {
     private String treatmentDetails; // تفاصيل العلاج
 
     @Column(nullable = false)
-    private Double amount; // قيمة المطالبة
+    private BigDecimal amount; // قيمة المطالبة
 
     @Column(nullable = false)
     private LocalDate serviceDate; // تاريخ تقديم الخدمة الطبية
